chore: rename all refs from ci to dev
This fixes the linting errors in CI.
