+++ .github/actions/call/action.yml
@@ -54,7 +54,7 @@ runs:
       if: inputs.dev-engine == 'true'
       with:
         go-version: "1.22"
-        cache-dependency-path: "ci/go.sum"
+        cache-dependency-path: "dev/go.sum"
 
     - name: Start dev dagger
       shell: bash
@@ -68,7 +68,7 @@ runs:
 
         # put env variables in /tmp/actions/call/local-envs instead of
         # $GITHUB_ENV to avoid leaking into parent workflow
-        (cd ci/mage; go run main.go -w ../.. engine:dev) | tee /tmp/actions/call/local-envs
+        (cd dev/mage; go run main.go -w ../.. engine:dev) | tee /tmp/actions/call/local-envs
 
         echo "::endgroup::"
       env:
