+++ CONTRIBUTING.md
@@ -9,7 +9,7 @@ on something that will not get merged.
 ## Building/Running/Testing
 
 For more detailed instructions on building, running and testing dagger locally,
-see the dagger [ci module](https://github.com/dagger/dagger/tree/main/ci).
+see the dagger [dev module](https://github.com/dagger/dagger/tree/main/dev).
 
 Working on dagger requires dagger to bootstrap it - you can install dagger
 using the instructions at [https://docs.dagger.io/install](https://docs.dagger.io/install).
@@ -195,7 +195,7 @@ The rest of this section gives some guidance on fixing these vulnerabilities whe
 
 ### Vulnerability in a Dagger binary
 
-If a vulnerability is reported in the Go stdlib, we'll want to upgrade the version of Go we use to build everything. As of this writing, this can be done by changing `GolangVersion` in `ci/consts/versions.go`.
+If a vulnerability is reported in the Go stdlib, we'll want to upgrade the version of Go we use to build everything. As of this writing, this can be done by changing `GolangVersion` in `dev/consts/versions.go`.
 
 Otherwise, if a vulnerability is reported in a Go dependency, you'll want to track down where the dependency is coming from.
 
@@ -207,9 +207,9 @@ This can become a bit complicated since it's possible for multiple versions of a
 
 ### Vulnerability in a third-party binary
 
-If a vulnerability is reported in the Go stdlib, we'll want to upgrade the version of Go we use to build everything. As of this writing, this can be done by changing `GolangVersion` in `ci/consts/versions.go`.
+If a vulnerability is reported in the Go stdlib, we'll want to upgrade the version of Go we use to build everything. As of this writing, this can be done by changing `GolangVersion` in `dev/consts/versions.go`.
 
-Otherwise, you'll want to check if the binary in question has a newer version with the vulnerability gone. The versions of these binaries are also controlled in `ci/consts/versions.go`.
+Otherwise, you'll want to check if the binary in question has a newer version with the vulnerability gone. The versions of these binaries are also controlled in `dev/consts/versions.go`.
 
 If there isn't a newer version to upgrade to, we'll be in a tougher spot and may need some combination of upgrading to a non-released commit, sending patches upstream or (as a worst-case fallback) patching it ourselves. Reach out to the Dagger team on Github or Discord if you're unsure how to best proceed.
