+++ dev/cli.go
@@ -6,8 +6,8 @@ import (
 
 	"golang.org/x/sync/errgroup"
 
-	"github.com/dagger/dagger/ci/build"
-	"github.com/dagger/dagger/ci/internal/dagger"
+	"github.com/dagger/dagger/dev/build"
+	"github.com/dagger/dagger/dev/internal/dagger"
 )
 
 type CLI struct {
