+++ dev/config.go
@@ -9,8 +9,8 @@ import (
 
 	"golang.org/x/exp/maps"
 
-	"github.com/dagger/dagger/ci/consts"
-	"github.com/dagger/dagger/ci/internal/dagger"
+	"github.com/dagger/dagger/dev/consts"
+	"github.com/dagger/dagger/dev/internal/dagger"
 )
 
 const (
