+++ dev/docs.go
@@ -5,8 +5,8 @@ import (
 
 	"golang.org/x/sync/errgroup"
 
-	"github.com/dagger/dagger/ci/internal/dagger"
-	"github.com/dagger/dagger/ci/util"
+	"github.com/dagger/dagger/dev/internal/dagger"
+	"github.com/dagger/dagger/dev/util"
 )
 
 type Docs struct {
