+++ dev/mage/util/dagger.go
@@ -17,7 +17,7 @@ func DaggerCall(ctx context.Context, args ...string) error {
 	}
 
 	cmd := exec.CommandContext(ctx, binary)
-	cmd.Args = append(cmd.Args, "call", "--source=.:default")
+	cmd.Args = append(cmd.Args, "call", "-m", "dev", "--source=.:default")
 	if path, err := hostDockerConfig(); err == nil {
 		cmd.Args = append(cmd.Args, "--docker-cfg=file:"+path)
 	}
