+++ dev/sdk_go.go
@@ -9,8 +9,8 @@ import (
 	"github.com/moby/buildkit/identity"
 	"golang.org/x/sync/errgroup"
 
-	"github.com/dagger/dagger/ci/consts"
-	"github.com/dagger/dagger/ci/util"
+	"github.com/dagger/dagger/dev/consts"
+	"github.com/dagger/dagger/dev/util"
 )
 
 type GoSDK struct {
