+++ dev/sdk_python.go
@@ -7,8 +7,8 @@ import (
 
 	"golang.org/x/sync/errgroup"
 
-	"github.com/dagger/dagger/ci/internal/dagger"
-	"github.com/dagger/dagger/ci/util"
+	"github.com/dagger/dagger/dev/internal/dagger"
+	"github.com/dagger/dagger/dev/util"
 )
 
 // TODO: use dev module (this is just the mage port)
