+++ dev/util/go.go
@@ -1,7 +1,7 @@
 package util
 
 import (
-	"github.com/dagger/dagger/ci/internal/dagger"
+	"github.com/dagger/dagger/dev/internal/dagger"
 )
 
 var dag = dagger.Connect()
